refactor: address daily quality findings for 2026-06-22
